refactor(TeacherProjectService): Move moveObjectUp/Down to array.ts
