+++ src/assets/wise5/common/array/array.ts
@@ -95,3 +95,19 @@ export function reduceByUniqueId(objArr: any[]): any[] {
   }
   return result;
 }
+
+export function moveObjectUp(objects: any[], index: number): void {
+  if (index !== 0) {
+    const object = objects[index];
+    objects.splice(index, 1);
+    objects.splice(index - 1, 0, object);
+  }
+}
+
+export function moveObjectDown(objects: any[], index: number): void {
+  if (index !== objects.length - 1) {
+    const object = objects[index];
+    objects.splice(index, 1);
+    objects.splice(index + 1, 0, object);
+  }
+}
